Verification tools in Python.
This means we now have code beyond just exposing data files in this repo.

This can easily be removed, but it also means the Bazel version needed is
explicitly set to something which was tested, so Bazelisk et al can operate
correctly. Removing it will likely happen, to reduce the scope of this repo
once again, possibly moving the changes into a branch or only into CI.

Python rules pull in at least 5.x, and remote build environment in an
upcoming commit pulls in at least 6.1.x (6.0.0 does not work).

6.5.0 is the chosen pre-Bzlmod version for running the tests.